refactor: movie card and add fallback image
