+++ src/components/MovieCard/MovieCard.jsx
@@ -1,9 +1,16 @@
 import { Card, Typography, CardMedia, CardContent, Box } from "@mui/material";
 import { PlayCircle } from "@mui/icons-material";
+import Fallback from "../../assets/broken-image.png";
+import styles from "./styles";
 import { truncateString } from "../../utils/stringUtils";
 
-const MovieCard = (props) => {
-  const { poster_path, genre_ids, vote_average, original_title, id } = props;
+const MovieCard = ({
+  poster_path,
+  genre_ids,
+  vote_average,
+  original_title,
+  id,
+}) => {
   const title = truncateString(original_title, 19);
   const genres = truncateString(genre_ids?.length && genre_ids.join(" "), 26);
 
@@ -11,58 +18,9 @@ const MovieCard = (props) => {
     console.log(id);
   };
 
-  const styles = {
-    container: {
-      position: "relative",
-      ":hover .MuiPaper-root": {
-        transition: "0.6s ease",
-        opacity: 0.3,
-      },
-    },
-    card: {
-      width: "200px",
-      height: "300px",
-      position: "relative",
-      bgcolor: "secondary.dark",
-    },
-    hoverLayer: {
-      width: "200px",
-      height: "300px",
-      position: "absolute",
-      left: 0,
-      top: 0,
-      opacity: 0,
-      textAlign: "center",
-      cursor: "pointer",
-      ":hover": {
-        transition: " .3s ease",
-        opacity: 1,
-      },
-    },
-    playIcon: {
-      color: "primary.main",
-      fontSize: "60px",
-      position: "absolute",
-      top: "calc(50% - 30px)",
-      left: "calc(50% - 30px)",
-    },
-    ratingBox: {
-      position: "absolute",
-      top: "20px",
-      left: "0px",
-      borderRadius: "5px",
-      height: "25px",
-      width: "40px",
-      bgcolor: "green",
-      textAlign: "center",
-    },
-    cardImage: {
-      width: "180px",
-      ml: 1,
-      mt: 1,
-      borderRadius: "10px",
-    },
-    genres: { display: "flex", textAlign: "center", columnGap: 1 },
+  const handleMediaFallback = (event) => {
+    event.target.style.objectFit = "contain";
+    return (event.target.src = Fallback);
   };
 
   return (
@@ -76,8 +34,10 @@ const MovieCard = (props) => {
         <CardMedia
           component="img"
           height="220px"
-          image={poster_path}
+          alt={title}
+          src={poster_path}
           sx={styles.cardImage}
+          onError={handleMediaFallback}
         />
         <CardContent sx={{ p: 1 }}>
           <Typography variant="h6" color="primary.light">
